+++ packages/terra-table/src/proptypes/validators.js
@@ -15,4 +15,21 @@ const validateRowHeaderIndex = (props) => {
   }
 };
 
-export default validateRowHeaderIndex;
+// eslint-disable-next-line consistent-return
+const validateAction = (props) => {
+  if (props.action) {
+    if (!props.action.label || props.action.label.length < 1) {
+      return new Error(ERRORS.ACTION_LABEL_MISSING);
+    }
+
+    if (!props.action.onClick) {
+      return new Error(ERRORS.ACTION_ONCLICK_MISSING);
+    }
+
+    if (props.action.onClick && typeof props.action.onClick !== 'function') {
+      return new Error(ERRORS.ACTION_ONCLICK_IS_NOT_A_FUNCTION);
+    }
+  }
+};
+
+export { validateRowHeaderIndex, validateAction };