+++ packages/terra-compact-interactive-list/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+* Fixed
+  * Added check for cell element existence before focusing.
+
+## 1.6.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
+## 1.5.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.4.0 - (February 16, 2024)
 
 * Changed