+++ packages/terra-form-validation/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.100.0 - (February 28, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 1.99.0 - (February 22, 2024)
 
 * Changed